refactor: static 리소스 경로 판단 로직 수정
